Completed one to one communication of one client chats with one admin only. Explanation of code: Problem now is if multiple admins are logged in and a user sends a message to the admin, the message get sent to ALL the admins which is not intended behaviour. We want one user to one admin communication.
We declare client variable and we use the active chats array and find all the client id that matches socketId inside this empty array.
If client is found, then target Admin id is client admin Id. Otherwise if, client is not found then get random admins. Next we create a get random function that finds random array of admins for chat. After that we can push to the activeChats array where client Id equals socketID.

Next, we update socket broadcast. Since socket broadcast sends it to all clients except the sender, but we want a particular admin to receive the client message. we add another method of broadcasting it TO a target Admin so only a particular admin receives message from a user. On the socket admin sends message to client, we broadcast admin message to a particular user.

In the AdminChatRoom Component, when socket emits message from admin, we sent id to socket User.
Recap of socket broadcasting works:
If a user starts a conversation:
we check that particular connection is in active chats array. If it is then assign to target admin Id the variable client admin Id. This means that the conversation continues from a previous instance. If client is not found, then this means it’s a new client and new conversation and we get a random admin that answers this client’s message.